Add files via upload
Project #2 – NetProbe Client-Server Edition
A. Overview and Specifications
In this project, we extend the NetProbe application developed in Project #1 into two applications, namely a client and a server application. 
(a)	The NetProbe Server will listen for incoming connections originating from a NetProbe Client, receiving the operating parameters (e.g., packet size, transmission rate, etc.) and then begin transmitting data to the NetProbe Client. Additionally, the NetProbe Server will be extended to become a concurrent server and thus, will be able to establish more than one NetProbe Sessions with multiple NetProbe Clients. The concurrent server model is to be implemented using multi-threading. No GUI is required for the NetProbe Server.
(b)	Both NetProbe Server and NetProbe Client are to be written such that they are portable for compilation and execution in both Linux and Windows, using a common set of source codes.