Add files via upload
Project #3 NetProbe Web Server
A. Overview
This project extends the NetProbe Server in Project #2 to support basic HTTP protocol processing such that the NetProbe Server can serve HTTP GET requests from HTTP clients (e.g., browser).
Only basic HTTP processing needs to be implemented:
- Only HTTP GET request needs to be supported;
- No need to support advanced HTTP options such as byte-range requests, persistent connection, etc.
The concurrent server is to be implemented using the same multi-threading model as in Project #2, using one-thread per connection. 
The threading model is to be implemented using two methods: (a) on-demand thread creation; and (b) thread-pool models. In model (a) a new thread is created for each incoming connection accepted, and destroyed after the connection ends. In model (b) a preconfigured fixed number of threads are pre-created and assigned to incoming connections. Upon connection completion the thread is returned to the pool rather than destroyed.