Add files via upload
Project #4  NetProbe Cloud Storage Application
A. Overview
This project integrates and extends the implementations in Project #2 and #3 into a cloud storage application. The cloud server extends the Web Server in Project #3 while the cloud client extends the NetProbe Client in Project #2 to make HTTP POST request to the cloud server.
The cloud client runs under Windows as a system service. It maintains a local directory configured by the user where all files in the directory are synchronized with the cloud server. The cloud server runs under Linux and maintains a local directory for where the files mirrored the ones in the user’s Windows host running the cloud client.
The application must implement the following functions:
(i)	Allow the user to specify the Windows directory for the local file. The local directory can be an existing directory or a new one.
(ii)	Whenever the cloud client is started it will establish connection with the cloud server and then synchronize the files by comparing the local files with the ones in the cloud server. If the local file is a newer version or a new one, then it will be uploaded to the cloud server. If the cloud server’s version is newer or a file does not exist in the local file system, then it will be downloaded back to the client Windows host.
(iii)	While the cloud client is running it will monitor the local and remote cloud storage by periodically waking up to check if there is any file updates, and if so, perform synchronization accordingly. The cloud client must implement proper file locking to avoid data corruption or race conditions.
(iv)	The cloud-storage application protocol is to be implemented on top of HTTP, where the cloud server will appear to be an HTTP server and the client will appear to be an HTTP client. To enable uploading of data to the server the application can make use of HTTP POST for data upload.
(v)	As HTTP is stateless the application will need to maintain state information using either HTTP Cookies or URL rewrites.